Cleanup build and clean scripts

+++ cleanup.sh
@@ -4,9 +4,12 @@ echo "Total containers: ${#names[*]}"
 
 for item in ${names[*]}
 do
-    echo "Cleaning container: $item"
-    sudo docker stop $(sudo docker ps | grep $item | awk '{print $1}')
-    sudo docker rm $(sudo docker ps -a | grep $item | awk '{print $1}')
-    sudo docker rmi $(sudo docker images | grep $item | awk '{print $3}')
+    RUNNING=$(sudo docker ps | grep $item | awk '{print $1}')
+    if [ ! -z $RUNNING ]; then
+        echo "Cleaning container: $RUNNING"
+        sudo docker stop $(sudo docker ps | grep $item | awk '{print $1}')
+        sudo docker rm $(sudo docker ps -a | grep $item | awk '{print $1}')
+        sudo docker rmi $(sudo docker images | grep $item | awk '{print $3}')
+    fi 
 done
